Fix Command equality
